+++ changelog.md
@@ -1,6 +1,7 @@
 - Custom AI: Number of improvements to collision avoidance logic. Ships are now less likely to bump into allies.
 - Custom AI: Ships now avoid mines.
 - Fixed log spam when calculations result in Not a Number. Suggested by Oddish.
+- Removed mod_info.json link to ASM library to avoid conflicts with other mods that depend on different ASM versions. The library is now loaded with a custom loader. Suggested by andylizi.  
 
 Version v2.2.1